+++ src/oic/oic/__init__.py
@@ -15,9 +15,11 @@ from typing import cast  # noqa - Used for MyPy
 from urllib.parse import parse_qs
 from urllib.parse import urlparse
 
+from jwkest import BadSyntax
 from jwkest import as_bytes
 from jwkest import jwe
 from jwkest import jws
+from jwkest import jwt
 from jwkest.jwe import JWE
 from requests import ConnectionError
 
@@ -1346,9 +1348,14 @@ class Client(oauth2.Client):
 
         headers = {"content-type": "application/json"}
         if registration_token is not None:
-            headers["Authorization"] = (
-                "Bearer " + b64encode(registration_token.encode()).decode()
-            )
+            try:
+                token = jwt.JWT()
+                token.unpack(registration_token)
+            except BadSyntax:
+                # no JWT
+                registration_token = b64encode(registration_token.encode()).decode()
+            finally:
+                headers["Authorization"] = "Bearer " + registration_token
 
         rsp = self.http_request(url, "POST", data=req.to_json(), headers=headers)