+++ CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on the [KeepAChangeLog] project.
 
 ## Unreleased
 
+### Added
+- [#719] Add support for JWT registration tokens
+
+[#719]: https://github.com/OpenIDC/pyoidc/pull/719
+
 ## 1.1.2 [2019-11-23]
 
 ### Fixed
@@ -16,7 +21,7 @@ The format is based on the [KeepAChangeLog] project.
 
 [#711]: https://github.com/OpenIDC/pyoidc/pull/711
 [#712]: https://github.com/OpenIDC/pyoidc/pull/712
-[#712]: https://github.com/OpenIDC/pyoidc/pull/717
+[#717]: https://github.com/OpenIDC/pyoidc/pull/717
 
 ## 1.1.1 [2019-11-04]